fix: support input float for duration range (#4447)
* support float

* slow analyzer

* sprintf float64

+++ modules/msp/apm/service/view/table/error_transaction.go
@@ -57,8 +57,8 @@ func (t *ErrorTransactionTableRow) GetCells() []*Cell {
 
 type ErrorTransactionTableBuilder struct {
 	*BaseBuildParams
-	MinDuration int64
-	MaxDuration int64
+	MinDuration float64
+	MaxDuration float64
 }
 
 func (t *ErrorTransactionTableBuilder) GetBaseBuildParams() *BaseBuildParams {
